% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{adjust_margin}
\alias{adjust_margin}
\title{Adjust the margin around plot}
\usage{
adjust_margin(right = 0, left = 0, top = 0, bottom = 0)
}
\arguments{
\item{right, left, top, bottom}{Number of pts to increase margin by.}
}
\description{
Increase \code{right} margin if you need more space for your x-axis label.
}
