% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_nsw.R
\docType{data}
\name{nsw_colours}
\alias{nsw_colours}
\alias{define_colour_theme}
\alias{col_nsw}
\title{Access NSW palette grids by rows and/or columns}
\usage{
nsw_colours

define_colour_theme(name, parent, colours)

col_nsw(
  hue,
  tone,
  variant = getOption("nswtheme.colour_theme", default = "base"),
  byrow = FALSE
)
}
\arguments{
\item{name}{name for the new colour theme.}

\item{parent}{name of the parent grid or theme, usually \code{"base"} or \code{"aboriginal"}.}

\item{colours}{character vector of colour column names.}

\item{hue}{name or index of the hue - see below.}

\item{tone}{name or index of the tone - see below.}

\item{variant}{name of palette variant.
Available options are: base, aboriginal, corporate, treasury.}

\item{byrow}{vary \code{tone} faster than \code{hue} if \code{TRUE}.}
}
\value{
\itemize{
\item for \code{col_nsw()} a named vector of colours,
\item for \code{define_colour_theme()} nothing: this is called for its side effects.
}
}
\description{
The NSW design system colours work in grids of colour columns and tonal rows.
\itemize{
\item \code{col_nsw()} allows accessing a colour grid like a matrix.
For ggplot colour palettes, you'll normally want \code{\link[=pal_nsw]{pal_nsw()}} instead.
\item \code{define_colour_theme()} defines a new colour theme that can be used
in any nswtheme function that accepts a \code{variant} parameter.
}
}
\details{
\code{hue} and \code{tone} work the same way as matrix indexing with \code{[} in that they
can be used to return single or multiple entries from the grid.
They can be character vectors, integer vectors, or logicals.

Anchor colours used to create the NSW colour palettes can also be used
stand-alone (e.g. \code{nsw_colours$blue_01} is \code{"#002664"}).
}
\section{Colour columns and tonal rows}{

The \code{"base"} variant supports:
\itemize{
\item \strong{hue}: greys, greens, teals, blues, purples, fuchsias, reds, oranges, yellows, browns
\item \strong{tone}: dark, normal, light, pale
}

The \code{"aboriginal"} variant supports:
\itemize{
\item \strong{hue}: reds, oranges, browns, yellows, greens, blues, purples, greys
\item \strong{tone}: dark, normal, light, pale
}

Colour themes support subsets of the hues from one of the main grids in a specific order.
These themes are built in:
\itemize{
\item \code{"treasury"}: teals, greys, oranges, greens
\item \code{"corporate"}: blues, reds, greys
}

The default variant can be specified globally with \code{options(nswtheme.colour_theme)}.

Unambiguous shortened forms are accepted, e.g. \code{pal_nsw(h = "red", v = "a")}.
}

\examples{
col_nsw(h = "blue", v = "aboriginal")
col_nsw(h = c("teal", "orange"), t = 1:2)
col_nsw(h = c("teal", "orange"), t = 1:2, byrow = TRUE)

}
\seealso{
\code{\link[=pal_nsw]{pal_nsw()}}
}
\keyword{datasets}
