% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.R
\name{coord_nsw}
\alias{coord_nsw}
\alias{coord_flip_nsw}
\alias{coord_polar_nsw}
\title{Coordinate transformations with corresponding changes to the ggplot theme}
\usage{
coord_flip_nsw(...)

coord_polar_nsw(...)
}
\arguments{
\item{...}{Optional arguments passed on to \code{\link[ggplot2]{coord_flip}()} or \code{\link[ggplot2]{coord_polar}()}.}
}
\description{
Wrapper functions for \code{coord_flip()} and \code{coord_polar()}. 
These functions transform your coordinate space; the \code{coord_*_nsw} equivalent does this \emph{AND} makes corresponding amendments to the NSW Government Brand theme.
\itemize{
 \item{flip: Flip cartesian coordinates so that horizontal becomes vertical, and vertical, horizontal.}
 \item{polar: The polar coordinate system is most commonly used for pie charts, which are a stacked bar chart in polar coordinates.}
}
}
\examples{
ggplot(data = mpg, mapping = aes(x = class, y = hwy)) +
  geom_boxplot() +
  coord_flip_nsw()
}
\references{
\url{https://nswgov.sharepoint.com/sites/inside-treasury/SitePages/Brand-Hub.aspx}
}
