% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\alias{forecast_shading}
\alias{forecast_line}
\title{Forecast line or shading}
\usage{
forecast_shading(
  date,
  label = "Forecast",
  facet = NULL,
  facet.variable = "category",
  fill = grey_02,
  alpha = 0.1
)

forecast_line(
  date,
  label = "Forecast",
  facet = NULL,
  facet.variable = "facet",
  linetype = 1,
  colour = grey_01
)
}
\arguments{
\item{date}{The latest historical date (before forecasts start).}

\item{label}{A character string to annotate the forecasts with. Defaults to "Forecast".}

\item{facet}{If used in a faceted ggplot, specify the facet(s) you want the \code{label} to appear on.}

\item{facet.variable}{If used in a faceted ggplot, the variable name used in your \code{facet_wrap()}. Defaults to "category", i.e. \code{facet_wrap(~category)}.}
}
\description{
Distinguish forecasts from historical data with either shading or a vertical line.
}
