% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{pal_interleave}
\alias{pal_interleave}
\alias{pal_c}
\alias{pal_stretch}
\alias{col_contrasting}
\alias{as_colour_vector}
\title{Construct palette variants}
\usage{
pal_interleave(...)

pal_c(...)

pal_stretch(pal)

col_contrasting(colour, light = "white", dark = "black")

as_colour_vector(x)
}
\arguments{
\item{...}{two or more vectors of colours.}

\item{pal}{palette object.}

\item{colour}{vector of colours.}

\item{light, dark}{colours to output when \code{colour} is dark or light respectively.}

\item{x}{a palette object, or a vector of colours to return unchanged.}
}
\value{
\itemize{
\item for \code{col_contrasting()} a vector of colours the same length as \code{colour},
\item for \code{as_colour_vector()} a vector of colours,
\item for \verb{pal_*()} a palette object.
}
}
\description{
Depending on the structure of your data you may wish to combine palettes
according to some pattern. These helpers may come in handy.
\itemize{
\item \code{pal_c()} concatenates multiple discrtee palettes.
\item \code{pal_interleave()} interleaves colours from multiple palettes of the same size.
It helps when your data are grouped and you need more flexibility than the colour grid.
\item \code{pal_stretch()} interpolates a palette into a new discrete palette.
It's useful for stretching the 4 tones into 5 or 6, at the expense of straying from
the NSW grid.
\item \code{col_contrasting()} chooses colours based on the given background colours.
It helps when drawing text on top of a mapped (i.e. variable) fill aesthetic
\item \code{as_colour_vector()} returns the colours of a discrete palette, which is
handy for passing them to functions that expect a colour vector.
}
}
