% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{ggsave_nsw}
\alias{ggsave_nsw}
\title{A wrapper for ggsave() with default dimensions}
\usage{
ggsave_nsw(
  filename,
  plot = last_plot(),
  width = 15.24,
  height = 9.98,
  units = "cm",
  dpi = 700,
  device = NULL,
  ...
)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2]{ggsave}()}.}
}
\description{
Save a ggplot (or other grid object) with useful defaults. This is the same as \code{\link[ggplot2]{ggsave}()}, but with the following features: 
\itemize{
 \item{Default size of 15.24cm x 9.98cm (width x height)}
 \item{Selects the best \code{device} option, when saving file as ".pdf" (\code{cairo_pdf}) or ".eps" (\code{cario_ps}) extensions.}
}
}
\references{
\url{https://nswgov.sharepoint.com/sites/inside-treasury/SitePages/Brand-Hub.aspx}
}
