% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_target.R
\name{inflation}
\alias{inflation}
\alias{inflation_target_lines}
\alias{inflation_target_shading}
\title{Inflation target lines or shading}
\usage{
inflation_target_lines(
  min = 2,
  max = 3,
  official = FALSE,
  start.date = NULL,
  linetype = "dashed",
  colour = grey_01,
  label = "RBA Inflation Target",
  label.x = "2017-08-01",
  label.y = 3,
  label.size = 7,
  label.colour = NULL,
  facet = NULL,
  facet.name = "facet",
  ...
)

inflation_target_shading(
  min = 2,
  max = 3,
  official = FALSE,
  start.date = NULL,
  fill = grey_01,
  alpha = 0.1,
  label = "RBA Inflation Target",
  label.x = "2017-08-01",
  label.y = 3,
  label.size = 7,
  label.colour = NULL,
  facet = NULL,
  facet.name = "facet",
  ...
)
}
\arguments{
\item{min, max}{Values to use in the plot. Defaults to 2 and 3, respectively.}

\item{official}{A logical determining whether to use the start date of only official, mandated targeting. Inflation targeting was adopted by the RBA in 1993, but was not mandated as an official target until 1996.}

\item{start.date}{A date, which if specified will override the default start date of inflation targeting. If used, \code{official} will be ignored.}

\item{colour, fill, alpha}{Colour/fill of lines/shading. Defaults to \code{grey_01}.}

\item{label}{A character string to annotate the target with. Defaults to "RBA Inflation Target".}

\item{label.x, label.y}{x- and y- coordinates for where to position the \code{label}.}

\item{label.size}{Size of \code{label} text (in pts). Defaults to 7, which is slightly smaller than other text (9).}

\item{label.colour}{Colour of \code{label} text. Defaults to be the same as \code{colour} or \code{fill}.}

\item{facet}{If used in a faceted ggplot, specify the facet(s) you want the inflation target to appear on.}

\item{facet.name}{The variable name used in your \code{facet_wrap()}. Defaults to "facet", i.e. \code{facet_wrap(~facet)}.}
}
\description{
Highlight the RBA inflation target with either shading or horizontal lines.
}
\examples{
# Data setup:
library(tidyverse)
library(readabs)
cpi <- read_abs(series_id = c("A3604506F", "A3604509L", "A3604503X"))
infl <- select(cpi, date, variable = series, value) |> 
  growth(4) |> 
  mutate(variable = str_remove(variable, "Index Numbers ;  "),
  variable = str_remove(variable, " ;  Australia ;")) |> 
  filter(year(date) >= 1990)

# Plot:
plot <- ggplot(infl, mapping = aes(x = date, y = value, col = variable)) +
geom_line() +
geom_hline(yintercept = 0) +
scale_colour_tsy(values = c(teal_01, fuchsia_02, fuchsia_03)) +
labs_tsy(y = "\%, Annual change", caption = "Source: ABS, NSW Treasury", title = "Headline and Underlying inflation") +
theme(legend.position = c(0.5,0.95))

plot + inflation_target_shading()
plot + inflation_target_lines()
}
\references{
Statement on the Conduct of Monetary Policy, Reserve Bank of Australia, September 1996, \url{https://www.rba.gov.au/monetary-policy/framework/stmt-conduct-mp-1-14081996.html}
}
