% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_xy_labels.R
\docType{data}
\name{label_nsw}
\alias{label_nsw}
\alias{en_dash}
\alias{em_dash}
\alias{label_comma_nsw}
\alias{fy}
\title{Label functions with sensible defaults}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
en_dash

em_dash

label_comma_nsw(
  accuracy = NULL,
  scale = 1,
  prefix = "",
  suffix = "",
  big.mark = ",",
  decimal.mark = ".",
  trim = TRUE,
  drop0trailing = TRUE,
  ...
)

fy(x, sep = en_dash)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[scales]{number}()}.}

\item{x}{Dates to translate into financial years.}
}
\description{
\itemize{
  \item{\code{label_comma_nsw} function behaves like \code{\link[scales]{label_comma}()}, but defaults to \code{drop0trailing = TRUE}.}
  \item{\code{fy} takes a vector of dates as inputs, and returns financial years as a \code{character} vector.}
  }
}
\keyword{datasets}
