% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{labs_nsw}
\alias{labs_nsw}
\title{A wrapper for labs() with useful defaults}
\usage{
labs_nsw(title = NULL, x = NULL, ..., tag = NULL, uppercase = FALSE)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[ggplot2]{labs}()}.}

\item{uppercase}{If \code{TRUE} (default), will capitalise \code{tag} and \code{title} text.}
}
\description{
This is the same as using \code{\link[ggplot2]{labs}()}, except \code{labs_nsw()}:
\itemize{
 \item{Puts the \code{tag} and \code{title} together on the same line.}
 \item{Defaults to \code{x = NULL}; since it is usually a date axis which doesn't need to be titled (\code{labs()} default is \code{x = waiver()}).}
 \item{Provides an option to capitalise the \code{tag} and \code{title}.}
}
}
\examples{
library(ggplot2)

ggplot(economics, aes(x = date, y = uempmed)) +
  geom_line() +
  labs_nsw(tag = "Figure 0.0:",
           title = "Median duration of unemployment",
           y = "Weeks")
}
