% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession_shading.R
\name{lockdown_shading}
\alias{lockdown_shading}
\title{Add a layer of COVID-19 lockdown shading}
\usage{
lockdown_shading(
  freq = "d",
  data = lockdowns_nsw,
  mapping = aes(x = NULL, y = NULL, colour = NULL, group = NULL, linetype = NULL, xmin =
    start, xmax = end, ymin = -Inf, ymax = Inf),
  fill = grey_01,
  alpha = 0.1,
  show.legend = FALSE,
  ...
)
}
\arguments{
\item{freq}{Frequency at which to show lockdown data. Defaults to "d". Options:\itemize{
\item{Daily [default]: \code{freq = "d"}}
\item{Weekly: \code{freq = "w"}}
\item{Monthly: \code{freq = "m"}}
\item{Quarterly: \code{freq = "q"}}
\item{Yearly: \code{freq = "y"}}
\item{Yearly (Financial): \code{freq = "fy"}}
}}

\item{...}{Arguments passed to \code{geom_rect()}.}
}
\description{
This is a wrapper for \code{\link[ggplot2]{geom_rect}()}, with convenient defaults for COVID-19 lockdown shading. Defaults to using the NSW lockdown dataset (\code{\link{lockdowns_nsw}}).
}
\seealso{
\code{\link{recession_shading}()}
}
