% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lockdowns}
\alias{lockdowns}
\alias{lockdowns_nsw}
\alias{lockdowns_vic}
\alias{lockdowns_qld}
\alias{lockdowns_wa}
\title{Lockdown start and end dates in Australia}
\format{
A data frame with 4 variables:
\describe{
  \item{start}{date the lockdown came into effect}
  \item{end}{date the lockdown was lifted}
  \item{days}{length of lockdown in days}
  \item{name}{Name of lockdown (descriptive only)}
  }

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 4 columns.
}
\usage{
lockdowns_nsw

lockdowns_vic

lockdowns_qld

lockdowns_wa
}
\description{
A dataset containing lockdown periods related to COVID-19, marked by start and end dates. Lockdown is assumed if at least one major city is in lockdown.\cr\cr
Lockdowns for:
\itemize{
  \item{New South Wales (not including the Northern Beaches lockdown)}
  \item{Victoria}
  \item{Queensland}
  \item{Western Australia}
}
}
\keyword{datasets}
