% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{nsw_colour}
\alias{nsw_colour}
\alias{grey_01}
\alias{grey_02}
\alias{grey_03}
\alias{grey_04}
\alias{green_01}
\alias{green_02}
\alias{green_03}
\alias{green_04}
\alias{teal_01}
\alias{teal_02}
\alias{teal_03}
\alias{teal_04}
\alias{blue_01}
\alias{blue_02}
\alias{blue_03}
\alias{blue_04}
\alias{purple_01}
\alias{purple_02}
\alias{purple_03}
\alias{purple_04}
\alias{fuchsia_01}
\alias{fuchsia_02}
\alias{fuchsia_03}
\alias{fuchsia_04}
\alias{red_01}
\alias{red_02}
\alias{red_03}
\alias{red_04}
\alias{orange_01}
\alias{orange_02}
\alias{orange_03}
\alias{orange_04}
\alias{yellow_01}
\alias{yellow_02}
\alias{yellow_03}
\alias{yellow_04}
\alias{brown_01}
\alias{brown_02}
\alias{brown_03}
\alias{brown_04}
\title{NSW Government colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
grey_01

grey_02

grey_03

grey_04

green_01

green_02

green_03

green_04

teal_01

teal_02

teal_03

teal_04

blue_01

blue_02

blue_03

blue_04

purple_01

purple_02

purple_03

purple_04

fuchsia_01

fuchsia_02

fuchsia_03

fuchsia_04

red_01

red_02

red_03

red_04

orange_01

orange_02

orange_03

orange_04

yellow_01

yellow_02

yellow_03

yellow_04

brown_01

brown_02

brown_03

brown_04
}
\description{
A vector of character strings containing HEX codes of all 10 colours in the \emph{NSW Treasury} palette.
}
\seealso{
\code{\link{nswgov_palette_simple}}
}
\keyword{datasets}
