% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nswtheme-package.R
\docType{package}
\name{nswtheme-package}
\alias{nswtheme}
\alias{nswtheme-package}
\title{Tools for the New South Wales Design System in R}
\description{
This package was inspired by the data viz work of Cara Thompson
(\url{https://github.com/cararthompson}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/digitalnsw/nswtheme}
  \item \url{https://digitalnsw.github.io/nswtheme/}
  \item Report bugs at \url{https://github.com/digitalnsw/nswtheme/issues}
}

}
\author{
\strong{Maintainer}: Zoe Baldwin \email{zoe.baldwin@health.nsw.gov.au} (\href{https://orcid.org/0000-0002-6010-9361}{ORCID})

Authors:
\itemize{
  \item Zoe Baldwin \email{zoe.baldwin@health.nsw.gov.au} (\href{https://orcid.org/0000-0002-6010-9361}{ORCID})
  \item Carl Suster \email{carl.suster@sydney.edu.au} (\href{https://orcid.org/0000-0001-7021-9380}{ORCID})
}

Other contributors:
\itemize{
  \item Shaun Nielsen \email{Shaun.Nielsen@health.nsw.gov.au} [contributor]
  \item Tyson Goddard \email{tyson.goddard@treasury.nsw.gov.au} [contributor]
}

}
