% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette}
\alias{palette}
\alias{nswgov_palette}
\title{NSW Government colour palette}
\format{
A data frame with 6 variables:
\describe{
  \item{name}{Colour name}
  \item{value}{HEX value}
  \item{colour_col}{Colour column of the palette}
  \item{tonal_row}{Tonal row of the palette - one of "01", "02", "03", or "04"}
  \item{corporate}{\code{TRUE} if colour is for corporate use}
  \item{core}{\code{TRUE} if colour is a core colour}
  }

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40 rows and 6 columns.
}
\usage{
nswgov_palette
}
\description{
A dataset containing the NSW Government palette.
}
\seealso{
\code{\link{nswgov_palette_simple}} is a \emph{named vector} equivalent
}
\keyword{datasets}
