% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palette_simple}
\alias{palette_simple}
\alias{nswgov_palette_simple}
\title{NSW Government colour palette (simple)}
\format{
An object of class \code{character} of length 40.
}
\usage{
nswgov_palette_simple
}
\description{
NSW Government palette in the form of a named vector.
}
\seealso{
\code{\link{nswgov_palette}} is a \emph{dataframe} equivalent with additional information
}
\keyword{datasets}
