% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_nswtheme.R
\name{reactable_nswtheme}
\alias{reactable_nswtheme}
\title{Theme for reactable interactive tables}
\usage{
reactable_nswtheme(
  colour = "blue_01",
  text_colour = colour,
  borderColor = colour,
  backgroundColor = "white",
  base_family = "Public Sans",
  title_family = "Public Sans",
  base_text_size = 10,
  stripedColor = NA,
  ...
)
}
\arguments{
\item{colour}{starting colour used only to set a common default for other
colour parameters.}

\item{text_colour}{colour of all text.}

\item{borderColor}{colour for the horizontal lines between rows.}

\item{backgroundColor}{background colour.}

\item{base_family}{font family for normal text.}

\item{title_family}{font family for title text.}

\item{base_text_size}{base text size in pt.
The header size is relative to the base text size.}

\item{stripedColor}{fill colour for alternate rows.
When \code{NA}, this is set to a blend of \code{borderColor} and \code{backgroundColor}.}

\item{...}{other parameters to pass to \code{\link[reactable:reactableTheme]{reactable::reactableTheme()}}.
For examples of what can be done with reactable, see
\url{https://glin.github.io/reactable/articles/cookbook/cookbook.html}.}
}
\value{
A reactable theme object
}
\description{
If using reactable, this helper constructs a theme.
}
\details{
Colours can be specified as named NSW colours as in \code{\link{nsw_colours}}.

As normal with HTML elements, you must make sure that the final document
loads the necessary fonts. See \code{vignette("nswtheme")} for instructions.
}
\examples{
library(reactable)

# The default theme
head(palmerpenguins::penguins, 10) |>
  reactable(theme = reactableTheme())

# Adding nswtheme style
head(palmerpenguins::penguins, 10) |>
  reactable(theme = reactable_nswtheme())

# More customised styling
head(palmerpenguins::penguins, 10) |>
  reactable(
    theme = reactable_nswtheme(
      colour = "blue_01",
      base_family = "Arial",
      text_colour = "black",
      backgroundColor = NULL
    ),
    striped = TRUE
  )

}
