% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession_shading.R
\name{recession_shading}
\alias{recession_shading}
\title{Add a layer of recession shading}
\usage{
recession_shading(
  data = recessions_aus,
  mapping = aes(x = NULL, y = NULL, colour = NULL, group = NULL, linetype = NULL, xmin =
    peak, xmax = trough, ymin = -Inf, ymax = Inf),
  fill = grey_01,
  alpha = 0.1,
  show.legend = FALSE,
  ...
)
}
\arguments{
\item{...}{Arguments passed to \code{geom_rect()}.}
}
\description{
This is a wrapper for \code{\link[ggplot2]{geom_rect}()}, with convenient defaults for recession shading. Defaults to recessions in Australia dataset (\code{\link{recessions_aus}}).
}
\seealso{
\code{\link{lockdown_shading}()}
}
