% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recessions}
\alias{recessions}
\alias{recessions_aus}
\alias{recessions_nsw}
\title{Recession peaks and troughs in Australia}
\format{
A data frame with 2 variables:
\describe{
  \item{peak}{date of peak, the quarter preceding any recession period}
  \item{trough}{date of trough, the quarter when the recessionary period reached its trough}
  }

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.
}
\source{
\url{https://www.abs.gov.au/statistics/economy/national-accounts/australian-national-accounts-national-income-expenditure-and-product/latest-release}
}
\usage{
recessions_aus

recessions_nsw
}
\description{
A dataset containing periods of economic recession, marked by the peak and trough. A recession is defined as two or more consecutive quarters of negative GDP growth. 
GDP sourced from ABS National Accounts (beginning December quarter 1959). For NSW, State Final Demand (SFD) is used (beginning September quarter 1985).
}
\keyword{datasets}
