% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_xy_labels.R
\name{scale_continuous_nsw}
\alias{scale_continuous_nsw}
\alias{scale_y_continuous_nsw}
\alias{scale_x_continuous_nsw}
\title{A wrapper for scale_*_continuous() that makes axis labelling easier}
\usage{
scale_y_continuous_nsw(
  limits,
  by,
  limits.secondary = NULL,
  by.secondary = NULL,
  name.secondary = NULL,
  expand.mult = c(0, 0),
  expand.add = c(0, 0),
  oob = scales::oob_keep,
  labels = label_comma_nsw(),
  labels.secondary = label_comma_nsw(),
  ...
)

scale_x_continuous_nsw(
  limits,
  by,
  limits.secondary = NULL,
  by.secondary = NULL,
  name.secondary = NULL,
  expand.mult = c(0, 0),
  expand.add = c(0, 0),
  oob = scales::oob_keep,
  labels = label_comma_nsw(),
  labels.secondary = label_comma_nsw(),
  ...
)
}
\arguments{
\item{limits}{A vector of values (of length 2), specifying [1] the minimum and [2] the maximum for the axis.}

\item{by}{Distance between ticks/labels.}

\item{limits.secondary, by.secondary, name.secondary}{Respective arguments on a secondary axis. If \code{by.secondary} not provided, will use the same magnitudes as the primary axis. Supply an axis title with \code{name.secondary}. Note: secondary axes are cosmetic only, so you will need to transform the variables you want on the axis beforehand (the function will calculate and return a warning with the coefficients you need for this).}

\item{expand.mult, expand.add}{Values passed to \code{\link[ggplot2]{expansion}()}. These can be used to expand the axis beyond the \code{limits} supplied, in one- or both- directions.}

\item{oob}{Function that handles limits outside of the scale limits (out of bounds). Defaults to \code{scales::\link[scales]{oob_keep}} to keep out of bounds data.}

\item{...}{Other arguments passed on to scale continuous function.}

\item{label.prefix}{A character string to prefix the label with (e.g. "$").}

\item{label.suffix}{A character string to suffix the label with (e.g. "\%").}
}
\description{
Specify your desired range for the axis, and formatting of labels.
}
\seealso{
\code{\link[ggplot2]{scale_x_continuous}()} and \code{\link[ggplot2]{scale_y_continuous}()}
}
