% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_xy_labels.R
\name{scale_date_nsw}
\alias{scale_date_nsw}
\alias{scale_x_date_nsw}
\alias{scale_y_date_nsw}
\title{A wrapper for scale_*_date() that makes axis labelling easier}
\usage{
scale_x_date_nsw(
  limits,
  by = "1 year",
  major = 1,
  format = "\%Y",
  FY = FALSE,
  label.between.ticks = FALSE,
  skip = 0,
  expand.mult = c(0, 0),
  expand.add = c(0, 0),
  oob = scales::oob_keep,
  ...
)

scale_y_date_nsw(
  limits,
  by = "1 year",
  major = 1,
  format = "\%Y",
  FY = FALSE,
  label.between.ticks = FALSE,
  skip = 0,
  expand.mult = c(0, 0),
  expand.add = c(0, 0),
  oob = scales::oob_keep,
  ...
)
}
\arguments{
\item{limits}{A vector of dates (of length 2), specifying [1] the minimum date and [2] the maximum date for the axis. Also accepts year values (numerics); these are converted to dates at the start [1] and end [2] of the given years, respectively (depends on \code{FY}).}

\item{by}{A string giving the distance between breaks/ticks (e.g. "2 weeks", or "10 years") - similar to date_breaks in . Defaults to '1 year'.}

\item{major}{Number of periods between date labels (Defaults to 1, i.e. label every period/year).}

\item{format}{\code{character} string of date format. Passed to \code{\link[base]{format.Date}()}. This argument is ignored if \code{FY = TRUE}. Example \code{2020-03-01}: \itemize{
\item{`format = "\%Y"` results in "2020"}
\item{`format = "\%y"` results in "20"}
\item{`format = "\%B"` results in "March"}
\item{`format = "\%d-\%b"` results in "01-Mar"}}}

\item{FY}{If \code{TRUE}, axis labels will be in the format "2011-12".}

\item{label.between.ticks}{If \code{TRUE}, will position the axis labels between tick marks - useful for showing the span of years more clearly.}

\item{skip}{How many periods to skip before date labels appear. Defaults to 0.}

\item{expand.mult, expand.add}{Values passed to \code{\link[ggplot2]{expansion}()}. These can be used to expand the axis beyond the \code{limits} supplied, in one- or both- directions.}

\item{oob}{Function that handles limits outside of the scale limits (out of bounds). Defaults to \code{scales::\link[scales]{oob_keep}} to keep out of bounds data.}

\item{...}{Other arguments passed on to scale date function.}
}
\description{
Specify your desired range of dates (or years), and formatting for \code{date} labels and ticks including whether or not to display in financial year notation.
}
\examples{
date_range <- c(lubridate::dmy('1-1-2000'), lubridate::dmy('31-12-2020'))

scale_x_date_nsw(xlim = date_range, label.between.ticks = TRUE, major = 2)
}
\seealso{
\code{\link[ggplot2]{scale_x_date}()} and \code{\link[ggplot2]{scale_y_date}()}
}
