% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_defaults.R
\name{set_defaults}
\alias{set_defaults}
\alias{set_defaults_nsw}
\title{Set NSW Treasury theme and palette as defaults}
\usage{
set_defaults_nsw(
  budget = FALSE,
  textbox = FALSE,
  base_size = NULL,
  base_family = "Public Sans",
  header_family = "Public Sans SemiBold",
  base_line_size = base_size/22,
  ink = NULL,
  paper = NULL,
  accent = NULL,
  ink_geom = teal_02
)
}
\arguments{
\item{budget}{a logical determining whether to use budget theme settings.}
}
\description{
Load fonts (including \emph{Public Sans} if installed) and set the default themes and colour scales. Run this function after loading packages.
}
\details{
The function will also attempt to set defaults for optional ggplot extensions such as ggpattern, ggrepel and geomtextpath. 
If these packages have not been loaded prior to setting defaults, the function won't be able to find the geoms - so it will skip them.
Therefore, if any of these packages are loaded later in your session, you'll need to set defaults again so the function can access the geoms.
}
\examples{
library(nswtrsy)

set_defaults_nsw()
}
