% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nsw.R
\name{theme_nsw}
\alias{theme_nsw}
\title{Theme plots using NSW colours and fonts}
\usage{
theme_nsw(
  base_size = 11,
  base_family = "Public Sans",
  header_family = "Public Sans",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ink = "black",
  paper = "white",
  geom_ink = "blue_01",
  accent = "blue_02",
  variant = getOption("nswtheme.colour_theme", default = "base"),
  show_grid_lines = TRUE,
  void = FALSE
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{header_family}{font family for titles and headers. The default, \code{NULL},
uses theme inheritance to set the font. This setting affects axis titles,
legend titles, the plot title and tag text.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{ink, paper, accent}{colour for foreground, background, and accented elements respectively.}

\item{geom_ink}{default ink colour used by geoms for points, lines and fills.}

\item{variant}{name of palette variant.
Available options are: base, aboriginal, corporate, treasury.}

\item{show_grid_lines}{whether to show grid lines.
If \code{FALSE}, all grid lines are removed but the axis text is retained.
Ignored when \code{void} is \code{TRUE}.}

\item{void}{whether to hide grid lines and axes.
If \code{TRUE}, all grid lines and axes are removed. This is useful when creating
pie/donut charts.}
}
\value{
ggplot theme specification to add to a plot
}
\description{
A 'ggplot2' theme compatible with the NSW design system.
It sets default colour scales, fonts, and some other styles.
}
\section{Compatibility}{

ggplot2 versions prior to 4.0.0 are missing some theme elements. If your'e
using an older version:
\itemize{
\item \code{accent} is accepted but unused,
\item geoms keep the ggplot2 defaults instead of \code{geom_ink},
\item NSW palettes need \code{\link[=scale_colour_nsw]{scale_colour_nsw()}} or \code{\link[=scale_fill_nsw]{scale_fill_nsw()}}.
}
}

\examples{
library(ggplot2)
theme_set(theme_nsw())

ggplot(palmerpenguins::penguins) +
  geom_point(aes(
    x = bill_length_mm,
    y = flipper_length_mm,
    colour = species,
    size = body_mass_g
  )) +
  labs(
    caption = "Data from {palmerpenguins}",
    x = "Bill length (mm)",
    y = "Flipper length (mm)",
    colour = "Species",
    size = "Body mass (g)"
  )

}
