% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip_css.R
\name{tooltip_css}
\alias{tooltip_css}
\title{Style interactive plot tooltips}
\usage{
tooltip_css(
  background_colour = "grey_01",
  text_colour = "off_white",
  font_family = "\\"Public Sans\\", Arial, sans",
  font_size = "11pt"
)
}
\arguments{
\item{background_colour}{tooltip background colour}

\item{text_colour}{tooltip text colour}

\item{font_family}{font family for text in tooltips}

\item{font_size}{font size for text in tooltips}
}
\value{
A character vector containing CSS rules
}
\description{
CSS code required to add nswtheme-styled tooltips to interactive graphs
created using ggiraph.
}
\details{
As normal with HTML elements, you must make sure that the final document
loads the necessary fonts. See \code{vignette("nswtheme")} for instructions.
}
\examples{
tooltip_css()

}
