% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_nsw}
\alias{theme_nsw}
\alias{theme_nsw_map}
\title{ggplot theme based on \strong{NSW Government} brand guidelines}
\usage{
theme_nsw(
  budget = FALSE,
  textbox = FALSE,
  base_size = 9,
  base_line_size = base_size/22,
  base_family = "Public Sans",
  header_family = "Public Sans SemiBold",
  ink = grey_01,
  paper = "white",
  accent = grey_04,
  legend.position = "inside",
  legend.position.inside = c(0.02, 0.98),
  legend.justification = c(0, 1),
  ...
)

theme_nsw_map(
  budget = FALSE,
  water_background = FALSE,
  textbox = FALSE,
  base_size = 9,
  base_line_size = base_size/22,
  base_family = "Public Sans",
  header_family = "Public Sans SemiBold",
  ink = grey_01,
  paper = "white",
  accent = grey_04,
  legend.position = c(0.97, 0.03),
  legend.justification = c(1, 0),
  ...
)
}
\arguments{
\item{legend.position}{The position of legends ("none", "left", "right", "bottom" or "top" for positioning legend outside plot) OR (two-element numeric vector for positioning legend inside plot). Defaults to \code{c(0.02,0.98)} (near the top-left corner of the plot area).}

\item{legend.justification}{anchor point for positioning legend inside plot ("center" or two-element numeric vector) or the justification according to the plot area when positioned outside the plot. Defaults to \code{c(0,1)} (anchored to top-left corner of the legend box).}

\item{...}{Arguments passed on to \code{\link[ggplot2]{theme}()}.}

\item{water_background}{For maps, logical determining whether to use a blue \code{panel.background} to represent the presence of the oceans.}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}
}
\description{
A theme object for the style of \emph{New South Wales Government}.
}
\examples{
library("ggplot2")

p <- ggplot(data = filter(economics_long, variable \%in\% c('uempmed', 'psavert'))) + 
  geom_line(mapping = aes(x = date, y = value, colour = variable)) +
  theme_nsw() +
  scale_colour_nsw()

# Facets:
p <- ggplot(data = mpg) +
  geom_jitter(mapping = aes(x = displ, y = hwy, colour = class)) +
  facet_wrap(~ manufacturer, nrow = 3)
  theme_nsw() +
  scale_colour_nsw()
}
\references{
\url{https://www.nsw.gov.au/sites/default/files/_static/nsw-gov-masterbrand-guidelines.pdf}
}
